requirements.txt
python-dotenv
Pillow
openpyxl
Jinja2